listController 기능 추가
