--- src/main/java/org/koreait/board/services/BoardList.java
@@ -1,5 +1,0 @@
-package org.koreait.board.services;
-
-public class BoardList {
-
-}
